changed attackComponent function to act on health values instead of character components
This will allow attacking with arbitrary damage values (weapon/spells),
and now the test suite uses QuickCheck to get an encompassing test for
the function.

+++ Component.hs
@@ -114,9 +114,9 @@ data Faction = Betuol | Dunteg | Blitzal
                deriving (Show , Read, Eq, Ord)
 type Reputation = (Faction, Int)
 data CharacterComponent = CharacterComponent
-    { health  :: Float
+    { health  :: Int
     , damage  :: Float
-    , mana    :: Float
+    , mana    :: Int
     , faction :: Faction
     , rep     :: [Reputation]
     } deriving Show
